Fix cublasLtMatmul failure (#167873)
This PR fixes a cublasLtMatmul failure seen in
`python test/test_ops.py TestCommonCUDA.test_out_warning_torch__scaled_mm_cuda`
caused by the use of `result_ptr` for matrix C. Using `nullptr` fixes this issue when `beta` is on host, which currently seems to be the only possible case, as the optional `alpha` parameter is not exposed to the user.

Using `result_ptr` for matrix C should be valid if Cdesc == DDesc or Cdesc is `nullptr`, but both of these cases cause `cublasLtMatmulAlgoGetHeuristic` to fail with `CUBLAS_STATUS_NOT_SUPPORTED`. The use of `nullptr` in `cublasLtMatmul` is a workaround until this failure can be resolved.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/167873
Approved by: https://github.com/slayton58, https://github.com/eqy