[opaque obj] Set type of FakeScriptObj to be OpaqueObj type (#167741)
When we create the FakeScriptObject we provide the name of the type of the FakeScriptObject. In the case of actual ScriptObjects this would be the CustomClassHolder class name, and for opaque objects this would've previously been a dummy name OpaqueTypeStr.

This PR changes the dummy name to be the actual opaque object type (addresses https://github.com/pytorch/pytorch/pull/163936#discussion_r2519860197).

Pull Request resolved: https://github.com/pytorch/pytorch/pull/167741
Approved by: https://github.com/zou3519
ghstack dependencies: #167739, #167740