Add GroupName NewType (#167552)
Use a NewType for `GroupName` instead of a raw string so that the typechecker has a chance to notice misuse - the only place we should be turning a raw string into a `GroupName` is where we create the ProcessGroup and assign the name in the first place. Everywhere else it should just be used as an opaque value.

While in there I added a few trivial types to some parameters in distributed_c10d.py as well.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/167552
Approved by: https://github.com/xmfan