[Inductor] MixOrderReduction support low precision reduction (#169978)
The issue does not appear for RMSNorm/LayerNorm since even if input is in low precision, RMSNorm will cast the inputs to FP32 first and do a down cast in the end. The reduction output is actually in FP32.

This PR fixes the issue if the reduction output is in low precision. Better to fix it to make MixOrderReduction a general feature.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/169978
Approved by: https://github.com/eellison