Symbolic Shapes: Ignorable fresh unbacked symbols (#169547)
When running fake meta impls we check that all the created SymInts are actually reachable from the returned values. We also potentially rebind the symbols we do find.

Currently if a meta needs to create symbols that it knows it won't return it we allow it to ignore them - but then they won't get recorded at all (and more importantly if we do return them they don't get rebound).

But we don't have a way to have a meta create symbols that it *might* return. So this PR adds a way to specify that some symbol is safe to not return but if we do return it then it should be processed normally.

I also looked into making all ignored symbols handled this way but there is some code that fails tests doing this - so it seemed safer to extend the behavior rather than modify the existing behavior.

This change is needed for compile-on-one-rank because the DTensor code which we use has a lot of generic helpers which may create an unbacked symint representing the rank and then not actually end up use it.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/169547
Approved by: https://github.com/bobrenjc93