[opaque_obj] Add nested value-type opaque obj support (#169845)
In the case of DTensorSpec, it contains multiple opaque objects like DeviceMesh and Placements. So when creating the repr, we need to get the list of nested types too to be added to the FX graph or inductor codegen.

The graph with nested value types would look something like:
```
def forward(self, arg0_1):
    process_nested_config = torch.ops._TestOpaqueObject.process_nested_config.default(arg0_1, NestedValueSize(size=SizeStore(size=3), config=ValueConfig(mode='square')));  arg0_1 = None
    return (process_nested_config,)
```

Pull Request resolved: https://github.com/pytorch/pytorch/pull/169845
Approved by: https://github.com/zou3519
ghstack dependencies: #167768, #169846