Update IS_THOR to SM 11.0 (#169719)
The `IS_THOR` check in torch/testing/_internal/common_cuda.py looks for a compute capability of 10.x, where x > 0. Thor is 11.0, so this check fails on Thor and gives a false positive on B300 and GB300 devices, which are 10.3.

Updated the check to look for 11.0 instead.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/169719
Approved by: https://github.com/eqy