[inductor][caching] simplify impls (#170220)
Summary:
Previously, when the interfaces were quite complex it made sense to abstract away the types that we put into the cache implementations. That way we shared complexity between interfaces and implementations.

Not that the interfaces are less complex, we can move this abstraction out of the implementations.

Mainly:
- cache impls now take str keys and bytes values
- except for the InMemoryCache, which takes Any values (with the assumption that Any is JSON-able)

This also has the added benefit of making InMemoryCache memory completely human readable (whether it is human readable in a way that actually makes sense is up to the user's encoding of results)

Test Plan: ```buck test fbcode//mode/opt caffe2/test/inductor:caching```

Differential Revision: D88979063

Pull Request resolved: https://github.com/pytorch/pytorch/pull/170220
Approved by: https://github.com/aorenste