support multi BLOCK_M for flex-decoding (#170343)
Flex-decoding previously is only active in VLLM in very narrow cases. We use flex-decoding if the workload can be processed with a single BLOCK_M: i.e. `BLOCK_M >= seq_len_q * G`, where G is the ratio between Q head and KV head. We use BLOCK_M as 16 in the vllm integration.

Take llama3 8B as an example, G is 4 here. That means we use flex-decoding only if `seq_len_q` <= 4. This is very restrictive and make flex-decoding being skipped most of the time.

The PR enhance the flex-decoding kernel to work for multi-BLOCK_M.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/170343
Approved by: https://github.com/drisspg