[xpu][feature]Fallbacks memory efficient attention to math attention on XPU (#166936)
Updates the logic for selecting the memory efficient attention backend on XPU devices. The main change is that instead of returning an error when the memory efficient attention backend is requested (which is not supported on XPU), the code now falls back to the math backend and provides a warning. Additionally, a typo in the function name has been fixed.

Backend selection improvements:

* Changed the logic in `select_sdp_backend_xpu` to fall back to the `math ` backend with a warning when memory efficient attention is requested on XPU, instead of raising an error.

Code quality improvements:

* Fixed a typo in the function name from `can_use_mem_efficien_attention` to `can_use_mem_efficient_attention`.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/166936
Approved by: https://github.com/guangyey, https://github.com/LuFinch, https://github.com/EikanWang

Co-authored-by: Yu, Guangye <106960996+guangyey@users.noreply.github.com>
Co-authored-by: Eikan Wang <eikan.wang@intel.com>