[flex_attention] adds support for low precision K/V inputs in compiled mode with GPU (#170486)
Allow Keys and Values to be in lower precision than Queries for memory efficiency.
K and V are automatically upcast to Q's dtype in the kernel before matmul operations.
In eager mode or compiled mode with CPU, error will be thrown with mixed dtypes.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/170486
Approved by: https://github.com/drisspg